[STOR-663] Forte: porting changes made to ClusterLock in forte 1.0 to forte 2.0.
Removed the sThreadKey map because we don't allow recursive locks, and therefore that map is not needed anymore
Added better logic for removing the mutex from the sMutexMap. Check to see that the mutex in the map is only referenced once, in which case we can remove it from the map.