forte: add ability to set the current state of a state machine
During unit testing of state machines, we currently have to
set up conditions appropriately to step through a state machine.
With this change, we can set a state in the state machine
directly, allowing us to more directly test state machine steps.

This check in also adds a compare operator for StateMachineLogger,
and the ability output a state history via the << operator. This
makes StateMachineLogger objects suitable for comparison in gtest
ASSERT macros.

Change-Id: I67c17bdfc49a36d9a319a42182f63ff620b8bdb7
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Tested-by: build
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: build