forte: ProcessManager fix for leaving behind zombie processes
When ProcessManagerImpl kicks off a process, it waits on it with
WNOHANG immediately after the fork. In some instances, this
happens before waitpid can be called. When waitpid is called
like this:
waitpid(-1, &child_status, WNOHANG);
it will wait on any children that have exited within our process
group. it appears that this call sometimes failed, but as new
processes are created, the old ones get wait'd on and reaped.

This patch waits on the specific child we care about until
it exits. If it is interrupted, it retries;

Change-Id: If3796135be341dba80c7890eca0c7fadec3343f6
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: build
Tested-by: build