forte ProcessManagerImpl: correctly handle failure of the daemon() call
If the call to daemon() fails, the procmon process must not be executed in the same
PID.  This causes the call to waitpid() in the parent (ProcessManagerImpl) to block
until the entire process completes, which is highly undesireable, and can block
the calling thread.

Failed calls to daemon() now will cause the child to exit with an error code,
which is correctly handled by the ProcessManagerImpl.

Change-Id: Ic14001666b8a3b110b79bf1c94fad4e64370608d
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Tested-by: build
Smoketests: build