forte: actually remove elements from mLogFiles when EndLogging(path) is called
the semantics of remove_if appear to pull forward elements
of the passed in item that you want to keep, leaving elements
at the end untouched. in the case that you use remove_if
without deleting the items returned via the forward iterator,
you will have a modified vector with multiple copies of
some elements. in this case, that would result in duplicate
log entries

Change-Id: I1c79b290a4f3a9a33afff99908eff9cb5a7b65fa
Reviewed-on: http://gerrit.lab.local:8080/1589
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>