Forte SM: implement inverted event handling STOR-2019
Forte SM: unit tests for new event delivery semantics

The state machine framework must not prevent inner regions and states
from receiving events which may be handled by enclosing regions.  This
is now implemented through an event queue at each region.  Events are
allowed to enter a state only if that state is stable, otherwise the
region queues the event until the state becomes stable.  States which are
not handled are dropped, and the initial delivery of an event is attempted
throughout the entire state machine recursively.

Change-Id: Ibcd0b861cf4515271120f1d6c04dc56193728e6c
Reviewed-on: http://gerrit.lab.local:8080/1817
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>