[STOR-2357] inject drive error into /var/log/messages
This change adds a test that will inject a message into syslog in a
form that will be picked up by the syslog parser as an error on
a drive. The test waits for the drive to be full removed as indicated
by the "driveNeedsReplacement" condition. After that is seen, scaled is
stopped, the error is cleared from the db and zookeeper, and scaled
is then started. The test then waits for drive related conditions
to clear.

Also:
  forte: added RescanAllHosts to SCSIUtil
  Renamed GetSlotDiskName to GetSlotDriveName
  Added drive tag to drive tests
  Updated DriveStateMachine tests, keeping as much as possible in tact
    but simplifying where checks were redundant
  Lots of ways to get drives in and out via lua
  Added mmdevdiscover and mmnsddiscover to re-add to kernel code
  tidied up output for scaled stable check

Change-Id: Ibfb6dda777d47fb1b4aa185781af4e69abb91b40
Reviewed-on: http://gerrit.lab.local:8080/2177
Tested-by: build
Smoketests: build
Reviewed-by: Mike McLaughlin <mmclaughlin@scalecomputing.com>
Reviewed-by: Nate Hardt <nate@scalecomputing.com>