forte: remove virtual destructor from AutoMutex classes
doing some performance testing of boost::mutex and forte::mutex, it
was found that the boost mutex was much faster. phil identified that
this was likely due to the virtual destructor we are getting from
forte::object. In a first pass, deriving directly from
boost::noncopyable also had a performance hit, so disabling copy
constructors was moved directly into the AutoMutex classes.
After the change, the forte mutex's speed is seen to be equivalent
to the boost classes.

An onbox test has been included, with some leeway for other
processes running. Running the tests on hardware have produced
fairly reliable results.

Change-Id: I1964a1ce39e4e769682789527bc555818892072d
Reviewed-on: http://gerrit.lab.local:8080/2268
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>