[STOR-3353] enhance PDUPeerSet to allow for easy setup in a cluster scenario
This change lays the groundwork for managing network connections inside of
the PDUPeer itself, and also allows for quick and easy implementation of
a PDUPeerSet in a cluster setup. It also keeps in place the ability to
simply create a PDUPeer based on a file descriptor in line with many
current usages.

Classes -
PDUPeerSetBuilder - this is the inteface to all of the classes and what
should be utilized to set up a clustered PDUPeerSet. The builder will put
together the classes needed to have a cluster of PDUPeers. The builder
will also be the interface to the PDUPeerSet, exposing methods EnqueuePDU,
SendPDU, and GetPeer which will pass on PDUs to one or all peers in
either a sync or async manner.

PDUPeerSet - PDUPeerSet lives on as the holder of the PDUPeers. It is
utilized internally by PDUPeerSetBuilder, and consumers should use the
inteface in PDUPeerSetBuilder.

PDUPeer - PDUPeer is a persistent representation of one end of a pdu
based conversation. It will keep a buffer of PDUs and attempt to send them
via it's endpoint. On failure, it will call an error callback
after the retry timeout has been exceeded.

PDUPeerEndpoint - PDUPeerEndpoint and it's subclasses represent the actual
connection between two PDUPeers. When creating a cluster of PDU peers via
PDUPeerSetBuilder, PDUPeerEndpoints that are ordered lower by their peers
in an IP/port ordering scheme will be created as Acceptors, while higher
peers will be created as Connectors.

This version works but it currently has a superfluous event queueu that needs
to be carefully managed. A new Thread type can be created that will work
on the queue that exists inside of PDUPeer. This will be done in a later
commit. Disabled, rusty unit tests that will be fixed in next
commit.

Change-Id: Iabd3f3c55f1f19d0e1d3ab1f4b9641e811f1f068
Reviewed-on: http://gerrit.lab.local:8080/2853
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>