[STOR-3745] forte:  Axe unused Context::Merge()
A polymorphic Merge() interface has a double-dispatch problem.[1]  Since
we don't call Merge() anyway, just remove it.  This is a prerequisite of
making Context a pure interface.

[1] Currently we have Context::Merge(const Context&).  If we're deriving
ContextImpl from Context, we'll have ContextImpl::Merge(const Context&),
where *this is a ContextImpl but the argument remains a plain Context --
at best we'd have to dynamic_cast to ContextImpl, throwing std::bad_cast
if the contexts were of different derived types.  This is ugly and fails
to fully implement the interface.  An alternative is to expose iteration
through the Context interface (either directly or by providing access to
the object map), but this is something not to consider until we actually
need it.

Change-Id: I2bd258bf8664a3154e5e961398533a1080d66a94
Reviewed-on: http://gerrit.lab.local:8080/2912
Tested-by: build
Smoketests: build
Reviewed-by: Josh Juran <jjuran@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Reviewed-on: http://gerrit.lab.local:8080/2949
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Smoketests: Nate Hardt <nate@scalecomputing.com>
Tested-by: Nate Hardt <nate@scalecomputing.com>