forte:Thread - abort if thread deletes itself
If the thread acquires a shared pointer to itself, (directly or
indirectly). It may end up being the owner of the last reference,
causing the destructor of the thread being called from it's own context.

Even though the object is destructed, the run method running inside of
the thread's context does not not of this. It is likely that it will try
to access data members of the object and cause corruption or hopefully a
segmentation fault.

Instead of allowing memory corruption to occur, we instead abort, to
allow for easier debugging, and identification of the troublesome owner.

Change-Id: I98db1b08f1b05bf9c134c7117ee17c5035b8a771
Reviewed-on: http://gerrit.lab.local:8080/3155
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Tested-by: build
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: build
Reviewed-on: http://gerrit.lab.local:8080/3240
Smoketests: Nate Hardt <nate@scalecomputing.com>
Tested-by: Nate Hardt <nate@scalecomputing.com>