forte: AsyncTask should hold mWaitLock until callback is complete
There is a small possibility where a callback could be called
on an object which has been deleted due to a Wait()er releasing
the sole shared_ptr to the task object after the wait condition
is broadcast.  By continuing to hold the wait lock until after
the callback is made, we assume that any Wait()er holding a
shared_ptr would not be able to release it due to needing to
re-acquire the mWaitLock upon waking from the Wait().

Change-Id: Id565781dc92f2bb8df1eac419e3a21ac63b0d1e0
Reviewed-on: http://gerrit.lab.local:8080/3396
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: Phil White <pwhite@scalecomputing.com>
Tested-by: Phil White <pwhite@scalecomputing.com>