forte: rename ClusterLock to InterprocessLock
The ClusterLock is actually a class that will hold a lock internal to a
process and/or external to a process via an advisory lock on a posix
file system. If the posix filesystem is a clustered posix filesystem,
this lock should also work across the cluster.

This commit changes the name, and also moves the test from being a unit
test to being an onbox test.

Change-Id: If7c689ec7a483235366ee1747c5f4e826acc2116
Reviewed-on: http://gerrit.lab.local:8080/3376
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: Phil White <pwhite@scalecomputing.com>
Tested-by: Phil White <pwhite@scalecomputing.com>