forte: do not abort on invalid PDUPeer
if a node is added and we are not informed about it we can be running in
a state where the cluster is consistent. aborting may still be the right
thing to do, but currently scaled will just do a scribe restart after
rolling out a new configuration file. this means we may briefly be in a
state where the cluster is not fully communicating, but it should be
remedied when the scaled impact runs.

Change-Id: Iaf2b2bb075e8261dd0c33eb2cc6ff43e705c4bd1
Reviewed-on: http://gerrit.lab.local:8080/3859
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>