forte:pdu: do not fill PDUPeer q for down node
In NetworkConnectorEndpoint, an event to check the connection was
enqueued any time a PDU failed to send. if the check connection event
failed, it would sleep for a 1 second while still holding the lock. this
would cause PDUs to continue to queue up for the down node. This patch
refactors the check connection logic to happen only on Send thread.

Removing that event also removed the last of the users of the
PDUWorkDispatcher in PDUPeerNetworkConnectorEndpoint, so it was removed
from that class.

The PDUWorkDispatcher then also become free of use, and was also
removed.

fixes [STOR-4794]

Change-Id: I63a216a81180cacf3cb46e5e0492d69c312e25a6
Reviewed-on: http://gerrit.lab.local:8080/4001
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>