forte: remove cause of spurious valgrind issues
there is an issue, still being investigated, where objects created
with make_shared and compiled with the c++0x flag create false
positives in valgrind. since log manager is used in everything, and
the objects it is creating with make shared happen very infrequently,
this commit changes LogManager to not use make_shared.

Change-Id: Ia2b51e8a48883ddf320dc55d1c8b84384ce64166
Reviewed-on: http://gerrit.lab.local:8080/4504
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>