[STOR-5397] forte ProcessManager: don't create thread until object is initialized
By creating the callback thread in the initializer list, the thread was being started prior
to the full construction of the ProcessManagerImpl object (specifically the mCallbackQueue
linked list, since it appears after the thread in the declaration).  This list was in an
undefined state when the thread was started, and there is a race where the thread can
access the list before it has been initialized.

While this change is an improvement, it is still not technically correct as the thread
is started prior to the constructor exiting.  A correct change would be the
implementation of the ThreadedObject interface for ProcessManager.

[STOR-5401] has been created for this issue.

Change-Id: I63de8146b387ae779bda520dc2c72ca82650303c
Reviewed-on: http://gerrit.lab.local:8080/4943
Reviewed-by: Josh Juran <jjuran@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>
Smoketests: Phil White <pwhite@scalecomputing.com>
Tested-by: Phil White <pwhite@scalecomputing.com>