forte:  Plug resource leaks in ThreadCondition
Explicitly make class instances non-copyable and move non-trivial member
functions out of the header.

Create stack-based class MonotonicClockThreadConditionAttr, derived from
ThreadConditionAttr.  Remove the condition attribute data member and use
a stack-based object in the constructor (which destroys the resource (a)
when it's no longer needed, (b) even in the event of exceptions).

Factor out ReportError() and append a line break to error messages.

(The Clock.h include can be downgraded to Exception.h here if we include
it as necessary downstream.)

Change-Id: I013fadadc2bfd363e1fd1b0798bfbe98ade16393
Reviewed-on: http://gerrit.lab.local:8080/5221
Tested-by: build
Smoketests: build
Reviewed-by: Josh Juran <jjuran@scalecomputing.com>
Reviewed-by: Nate Hardt <nate@scalecomputing.com>