[STOR-6706] Fix IOManagerUnitTest
The WaitForAllRequestsToComplete test had a race condition.  The IOManager may
make the callback for the last IO request in the queue *after* signalling
the condition indicating the queue is empty.  This means after the function
returns, we may still have to wait shortly for our count to be updated.

Change-Id: Ie08d51130a066dcb04d83f8d91ed7a895235b702
Reviewed-on: http://gerrit.lab.local:8080/6968
Tested-by: build
Smoketests: build
Reviewed-by: Nate Hardt <nate@scalecomputing.com>
Reviewed-by: Phil White <pwhite@scalecomputing.com>